+++ src/browser_stream.ts
@@ -1,4 +1,5 @@
 import { resolve } from "node:path"
+import { relative } from "node:path/posix"
 
 import { normalizeCwd } from "./normalizeCwd.js"
 import { opendir } from "./opendir.js"
@@ -19,7 +20,7 @@ export function scanStream(options: ScanOptions & { fs: FsAdapter; cwd: string }
 	const {
 		target,
 		cwd,
-		within: select = ".",
+		within = ".",
 		invert = false,
 		depth: maxDepth = Infinity,
 		signal = null,
@@ -44,7 +45,7 @@ export function scanStream(options: ScanOptions & { fs: FsAdapter; cwd: string }
 
 	const scanOptions: Required<ScanOptions> = {
 		cwd: normalCwd,
-		within: select,
+		within,
 		depth: maxDepth,
 		fastDepth,
 		fastInternal,
@@ -54,14 +55,16 @@ export function scanStream(options: ScanOptions & { fs: FsAdapter; cwd: string }
 		target,
 	}
 
-	const result = opendir(fs, resolve(normalCwd, select), (entry) =>
-		walkIncludes({
+	const result = opendir(fs, normalizeCwd(resolve(normalCwd, within)), (entry) => {
+		const path = relative(normalCwd, normalizeCwd(entry.parentPath) + "/" + entry.name)
+		return walkIncludes({
+			path,
 			entry,
 			ctx,
 			stream,
 			scanOptions,
-		}),
-	)
+		})
+	})
 
 	void (async (): Promise<void> => {
 		await result